Fade navbar-handle in/out during TRANSITION_HANDLE_FADE
The existing code relied on the INSTANT / FINAL_FRAME interpolators,
however these did not work as expected for sub-animations with shorter durations,
since the INSTANT / FINAL_FRAME interpolator applies to the sub-animations durations

This caused the stashed handle to jumpcut in/out, while the taskbar icons were
actually visible in the area of the navbar.

This is fixed to polish the lock/unlock animation (b/275319714), but happenes to fix b/276941963 as well.

Before: http://shortn/_uxw2oe7QeA
After: http://shortn/_yMY8IQIM60

Bug: 275319714
Bug: 276941963
Fixed: 275319714
Fixed: 276941963

Test: manual (http://shortn/_rnGNSkRH3w), tapl tests
Change-Id: I79e431e9ee21ede069864ee217bb343f0c32dfc7