Add spacing to gesture nav edu for large fonts.
For devices with smaller screens and the largest fonts and display
sizes, the end screen of the gesture navigation tutorial would cut off
the `Done` button at the bottom. This adds margins so that the button
remains visible on the screen.

Screenshot of Pixel 5 with the largest font and display:
https://screenshot.googleplex.com/AU5rYYBisZ9Azbr.png

Fix: 275540512
Test: Went through the tutorial with default and largest font sizes on both smaller devices and tablets to ensure the button is always visible.
Change-Id: I66baf9b9044211550395bb3823797127de5e1f99