Fix single frame flicker for views reacting to keyboard insets.
This includes Folders, bottom apps searchbar, work apps button.

The issue is that the view was being drawn between onStart and
onProgress. Before onStart is called, the final keyboard insets
are dispatched to determine the view's final state, so this draw
captured this otherwise invisible state. Normally, onProgress
would be called first, effectively resetting the view; but for
example, the transition from closing the apps drawer forces a
draw prematurely.

Fix: 265830720
Test: Manual
Flag: N/A, but improves the ENABLE_FLOATING_SEARCH_BAR flag.
Change-Id: I818cb75628dec6dcc69e5f4b8ae554c2245dceca