Remove unused Assistant gesture tutorial and add error checking
The Assistant gesture tutorial is unused in both the current and
redesigned gesture nav tutorials and can be removed. Left in, its
detections of gestures are leading to bugs with the tutorial.

Additionally, this change adds in error checking for side swipes in the
overview tutorial.

Flag: ENABLE_NEW_GESTURE_NAV_TUTORIAL
Fix: 280115831
Fix: 261239858
Fix: 278926113
Test: Manually went through both versions of gesture tutorial on
handheld, foldable, and tablet devices

Change-Id: Ie481d92ede2c2ab16ce2b3d1f729f6cb88f1b4c0