Un-Split TaplTestsTaskbar
Before we use TaskbarModeSwitch but had a drawback where we would not know which variant of taskbar the test failed for. So, we split the taskbar tests into transient and persistent for better error logs and debugging.
But, now we can use Parameterized Test Runner to achieve the same and do not have to duplicate test code for testing two different variants of Taskbar.

There can be instances where we only want to test certain taskbar variant only, we have separated files for that.

Test: Presubmit
Bug: 267648422
Flag: not needed
Change-Id: I4322498f13d257a5f1793b008f45b09ac9b9f3ff