Fix UI issues of fake task view in gesture nav tutorial.
This change fixes the bug where the view for the fake task in the overview gesture was not being reset correctly. This was leading to UI issues such as lack of rounded corners and incorrect animations in the gesture navigation tutorial.

Flag: ENABLE_NEW_GESTURE_NAV_TUTORIAL
Fix: 281121451
Fix: 281121699
Fix: 281785378
Test: Manually went through the tutorial on handheld, foldable, and
tablet devices

Change-Id: I424a471013d2b8667724f790618d142ba7ae9a7b