Migrate the insetsRoundedCornerFrame to InsetsFrameProvider
The server side has moved the InsetsRoundedCornerFrame into
InsetsFrameProvider as it is actually using a given type of insets frame
to calculate the new rounded corner information and moving it into the
corresponding provider would make it cleaner.

To make sure it can be updated and accessed properly,
TaskbarInsetsController is now holding the nav bar insets provider to
make the update. It will be used always even if the height has been
updated. Other necessary changes are made to reflect the change.

Test: DisplayPolicyInsetsTests
Test: m
Bug: 274724339
Change-Id: Iec6804a01bafbbaed8591694490f29c97ea7e7fc