Fix TaskView bind and orientation state called multiple times
DWB toast initialization is being called 3 times for every TaskView. Twice during the TaskView bind (setOrientationState and bind), and for every TaskView loaded in applyLoadPlan the bind and the setOrientationState are invoked. This CL removes the excessive calls in the TaskView bind and applyLoadPlan.
The DWB for the focused task view is still being called twice due to orientation update for the focused task.

Bug: 281980412
Flag: N/A
Test: atest DigitalWellBeingToastTest
Change-Id: Id301b53b9cb0dfc23a81debec67a6de9a35485b9