Use LockedUserState in TouchInteractionService.
This change was previously attempted, but failed because of a tricky issue where the LockedUserState singleton object was getting permanently set to a mock context in LockedUserStateTest, and then was failing TaplTests because isUserUnlocked was always false.

This fixes that by avoiding using the singleton LockedUserState instance in the unit tests.

Bug: 251502424
Test: Compilation threw no errors and user unlock behavior worked
correctly. Post-submit was tested and this CL passed all previously
failing tests: https://android-build.googleplex.com/builds/abtd/run/L12900000960898179

Change-Id: I045c9f2558a6bdacb4bfa029fbf6a07c3c190fe7