Match displayed instruction with voiceover text.
When users are using TalkBack, the text being read aloud did not match
the text displayed on the screen. This change fixes that in accordance
to the accessibility testing principles by changing the text spoken
aloud.

Flag: ENABLE_NEW_GESTURE_NAV_TUTORIAL
Fix: 274996517
Test: Manually went through both versions of the tutorial using talkback
and without using it. Ensured the spoken text matched the displayed
text

Change-Id: I896f2c09bd220615722777c73935f70c946715b2