Make sure FLAG_IN_APP is set when launching apps from keyguard.
This flag was not being set in the case where you lock from launcher and
then unlock by tapping on launcheable items from keyguard such as home
controls, notifications, smartspace, etc. This change makes sure we
double check if we are unlocking to launcher or not and then setting
FLAG_IN_APP accordingly.

Test: manual
Fix: 284314711
Change-Id: If46697566d252c649a00a7b3d728a14789dc6aed