Updating content description of taskview after latest update
Description: we wait for icon updates to receive the title of the task. However we still set the description immediately upon digitalwellbeingtoast creation.
this creates a null set for contentDescription resulting in incorrect talkbalk. This will listen for icon updates and will reiniitialize the toast ensuring
taskview is set properly.

Test: built and tested locally with talk back ensuring the titles are being set correctly

Bug: 280908761
Change-Id: I946e7190114cbdccb75883965b1fdf85ed6ad17a