Small tweak to floating search on phone.
When you swipe down from all apps back home, the search bar was
expanding due to the size being slightly larger when resting on
home. However this looks a bit off with the search protection,
so created this special case to keep it at the all apps size
during this transition.

Bug: 261866704
Test: Manual on tablet phone and foldable.
Flag: ENABLE_FLOATING_SEARCH_BAR
Change-Id: If9399224d0c456629d094139eef7f4d1b2c3bbc9