Reparent PiP overlay under task as soon as possible
For enter PiP animations started from Launcher, we send the signal to
WMShell to reparent the overlay under task as soon as possible.

This avoids flicker issue caused by race condition of cleaning up the
transition leash before the reparenting was done, which made the issue
of overlay being not attached to anything for few frames.

Bug: 281711939
Test: winscope trace from comment#28, leash follows task
Test: after http://recall/-/ekEuGtt9d9HWqkUtAzpHx8/ewrarENSYYOmUCa0b3fdtp

Change-Id: Ia16afd49a94d9ba8456287c2cdffb2a84e5a2bc7