Revert "Fix padding/margin for widgets in widgetpicker"
This reverts commit 3cc8de12682ce7f68bc17afc09b01ea6702b5b87.

Reason for revert: when adding this extra margin to widget_cell it has the unintended effect of making text width available for the widgets shorter than it originally was (E.g Dynamic Colors becomes Dynamic Co...). Adding padding to widgets_table_container.xml instead seems to be a better solution to this issue.

Change-Id: I90b850a43aca573a26131a96d5d8f5fbde8b5f6e