Ensure the overflow is added first to the bubble bar
Previously there was a race between initializing the bubble bar
which creates the overflow and sending an initial bubble state
update event from WMShell. If the event is processed first then
bubbles are added before the overflow.

This change ensures that the overflow is created if needed when
the initial event is sent from WMShell.

Fixes: 288414998
Test: - manual
        - Enable 3 button navigation
        - Create some bubbles
        - Switch to gesture navigation
        - Overflow should be added first (below other bubbles)
Change-Id: Ie001c76a3f587d95021f4b7202492b3071706fd0