Prevent exception when quick switching between two split pairs
When switching in between two split pairs just quick enough, it is
possible to send the second entering split transition while it is
animating the first entering split transition which is merged to a
recents-during-split transition. The split parents might not be
collected into the second transition because the split parents are
already visible at that time, and there is no recents transition to
merge to. So updates to not throwing when there is no split parent when
composing a recents-to-split animation.

Bug: 236226779
Test: repro steps of the bug and the Launcher won't throw
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:97eb471f2efda1eccce7eba8c68301b2fa34c5e7)
Merged-In: I3a595722721186e8de7d60c9fb8c099ec799804a
Change-Id: I3a595722721186e8de7d60c9fb8c099ec799804a