Don't show the bubble bar when it is empty
Sometimes the bubble bar shows when the launcher process starts up
even when there are no bubbles.

bubblesShowingOnHome is set to true when the launcher transitions into
Home, but that state also triggers a call to showBubbleBar() which will
eventually set the visibility on the bar.
During initialization we also call setHiddenForNoBubbles to hide the bubble
bar if it's empty, but these events happen separately so the bar may or
may not be hidden.

This change avoids calling showBubbleBar when setBubblesShowingOnHome is
called and the bar is invisible. Instead the bubble bar will become visible
once bubbles are added.

Fixes: 290608658
Test: manual
       - Toggle bubbles off and back on in settings
       - Restart launcher
       - Observe the bubble bar is not displayed
Change-Id: I30640764d7b6eae64e1c5b5c687760465bf9a7b1