Check that the bubble bar has bubbles before showing it
ag/24024252 relied on the bubble bar being invisible to indicate that
the bar is empty. But that assumption was wrong. The bubble bar gets
hidden when the device is locked. The result is that unlocking the
device leaves the bubble bar stashed.

This change actually verifies whether the bubble bar is empty before
showing it.

Fixes: 290608658
Test: Manual:
      - Have some bubbles in the bar
      - Lock the device and unlock
      - Observe that the bubble bar is unstashed
Change-Id: I4ca5c10e0cec57d3c6b4e056e06ebcf04aa1571a