Fix order invariance when checking for GroupedTaskViews
This patch makes it so that Overview will return the correct GroupedTaskView even when the set of taskIds passed in is in the opposite order.

Previously, we didn't account for order variation when checking for a GroupedTaskView, which would cause bugs when the GroupedTaskView had had its ordering swapped.

Fixed by modifying getTaskViewByTaskIds() to return true even if the order of ids was swapped.

Fixes: 290135090
Test: Manual
Change-Id: I2d94fadfee3cd478801b25cdbc699295179c8bd1