Set the stash handle bounds deterministically
The bounds of the bubble stash handle is calculated based on the bounds of
the bubble bar. The problem is that the bubble bar could be invisible, which
results in incorrect bounds set for the handle.

This change sets the bounds for the handle based on the width of the screen.

Fixes: 290992144
Test: Manual
        - Remove bubbles
        - Restart Launcher
        - Open the bubble test app and double tap on a chat
        - Observe that the handle animates in
Change-Id: Ida260014d59b88387de010891c18057f3b091e93