Disable FOLDABLE_SINGLE_PAGE
The functionality should go back to the same as with phones.
There shouldn't be issues with the reorder or similar behavior since we
are switching form using the MultipageCellLayout to the regular
CellLayout.

The things we need to pay attention to is the the behavior of having two
panels like adding the right number of panels when loading (folding, unfolding and rotating).

Bug: 291822492
Change-Id: I903873e32f35c5ee9e0f3da8581a37d4087d021f
Test: ReorderWidgets
Meged-In: I903873e32f35c5ee9e0f3da8581a37d4087d021f