Fix nav buttons in task bar disppearing when unfolding from landscape small screen
Previously we reuse layout params when we switch to a new layout, but not all of the fields we reuse are reset. Due to the variety of fields we update in different layouts and the fact that it's relatively cheap to instantiate a new layout params object, let's start fresh every time.

Fixes: 295057715
Bug: 288311605
Test: put the device to landscape in small screen and then unfold, make sure that the nav buttons show in the task bar
Change-Id: I8e600c2553436f563dd4509a83cc46ce47c0114d