Fix work paused text padding.
This is causing an issue for keyboard up by default users who can't see the button due to the keyboard blocking it.
The reason this is happening is because there is a layoutMargin of 40dp that is on the work_apps_paused_title text.
But it should be 16dp from the tabs and the and work_apps_paused_title text. Due to the floatingHeaderView having 8dp already,
just add another 8dp padding to the workPausedCard for the 16dp distance between the tabs and the text.

bug: 294948106
flag: n/a
test: manual -photos:
before:https://screenshot.googleplex.com/8awNqhWLkW3mF2w
after: https://screenshot.googleplex.com/8qUJAwH5Qzd6d3L

Change-Id: I4468e2fec6f05919323e27b1c4296d0239b2d044